add styles for sign up
