+++ src/Components/App/App.js
@@ -5,7 +5,7 @@ import { url } from '../../apiURL';
 import MovieContainer from '../../containers/MovieContainer/MovieContainer';
 import { Redirect, Route } from 'react-router-dom';
 import SignIn from '../../containers/SignIn/SignIn';
-import SignUp from '../SignUp/SignUp';
+import SignUp from '../../containers/SignUp/SignUp';
 
 export class App extends Component {
