+++ src/Main.scss
@@ -1,3 +1,4 @@
 @import './Components/App/App';
 @import './containers/MovieContainer/MovieContainer';
 @import './containers/SignIn/SignIn';
+@import './containers/SignUp/SignUp';
